fix .deb specification

+++ 04_install_prerequisites.sh
@@ -10,15 +10,25 @@ check_env_var CROSS_COMPILER_RELATIVE_URL
 set -ex
 
 # install libssl-dev and libz-dev
-apt-get update && apt-get install -y libssl-dev libz-dev
+apt-get update && apt-get install -y libssl-dev zlib1g-dev
+
+# install cURL
+apt-get update && apt-get install -y curl
 
 # install CMake 3.16
-apt-get update && apt-get install -y curl
-curl -L -o cmake-linux.sh https://github.com/Kitware/CMake/releases/download/v3.16.1/cmake-3.16.1-Linux-x86_64.sh
-sh cmake-linux.sh -- --skip-license --prefix=/usr
+if [ ! -f /usr/bin/cmake ]; then
+    curl -L -o cmake-linux.sh https://github.com/Kitware/CMake/releases/download/v3.16.1/cmake-3.16.1-Linux-x86_64.sh
+    sudo sh cmake-linux.sh -- --skip-license --prefix=/usr
+fi
 
 # install cross-compiler for Raspberry Pi
 # (https://github.com/abhiTronix/raspberry-pi-cross-compilers)
 mkdir -p ${CROSS_COMPILER_ROOT}
-curl -L -o ${CROSS_COMPILER_PACKAGE_FILE} ${CROSS_COMPILER_BASE_URL}/${CROSS_COMPILER_RELATIVE_URL}
+
+if [ ! -f ${CROSS_COMPILER_PACKAGE_FILE} ]; then
+    curl -L -o ${CROSS_COMPILER_PACKAGE_FILE} ${CROSS_COMPILER_BASE_URL}/${CROSS_COMPILER_RELATIVE_URL}
+fi
 tar -xzf ${CROSS_COMPILER_PACKAGE_FILE} -C ${CROSS_COMPILER_ROOT} --strip-components 1
+
+# install prerequisite build tools and dev packages
+apt-get install make g++
